Changed error status and message of api

+++ src/app/application.rb
@@ -6,10 +6,6 @@ require 'sinatra/reloader'
 require File.expand_path('../../lib/validator/validator.rb', __FILE__)
 require File.expand_path('../../lib/validator/auto_annotation.rb', __FILE__)
 
-#require_relative "../../ddbj_validator/src/biosample_validator/biosample_validator.rb"
-#require_relative  "/" + File.dirname(__FILE__) + "../../../ddbj_validator/src/biosample_validator/biosample_validator.rb" #.rb" # + "../../../src/biosample_validator/biosample_validator.rb"
-#require_relative  "../validator/biosample_validator/biosample_validator" #.rb" # + "../../../src/biosample_validator/biosample_validator.rb"
- 
 module DDBJValidator
   class Application < Sinatra::Base
     setting = YAML.load(File.read(File.dirname(__FILE__) + "/../conf/validator.yml"))
@@ -25,6 +21,10 @@ module DDBJValidator
 
     configure :development do
       register Sinatra::Reloader
+    end
+
+    before do
+      response.headers["Access-Control-Allow-Origin"] = "*"
     end
 
     get '/api/client/index' do
@@ -61,6 +61,7 @@ module DDBJValidator
     post '/api/:version/validation' do |version|
       unless version == @@latest_version #バージョンが最新でなければ400　本当は転送したい
         status 400
+        body "invalid version. latest version is '#{@@latest_version}'"
       else
         #組み合わせが成功したものだけ保存しチェック
         if valid_file_combination?
@@ -94,7 +95,6 @@ module DDBJValidator
             end
           }
 
-          response.headers["Access-Control-Allow-Origin"] = "*"
           content_type :json
           { uuid: uuid }.to_json
         else #file 組み合わせエラー
@@ -106,44 +106,53 @@ module DDBJValidator
     get '/api/:version/validation/:uuid' do |version, uuid|
       unless version == @@latest_version #バージョンが最新でなければ400　本当は転送したい
         status 400
+        body "invalid version. latest version is '#{@@latest_version}'"
       else
         save_dir = "#{@@data_dir}/#{uuid[0..1]}/#{uuid}"
         output_file_path = "#{save_dir}/result.json"
-        #ファイルがなければ400番?
-        result_json = JSON.parse(File.open(output_file_path).read)
-        response.headers["Access-Control-Allow-Origin"] = "*"
-        content_type :json
-        result_json.to_json
+        if File.exist?(status_file_path)
+          result_json = JSON.parse(File.open(output_file_path).read)
+          content_type :json
+          result_json.to_json
+        else
+          status 400
+          body "Invalid uuid"
+        end
       end
     end
 
     get '/api/:version/validation/:uuid/status' do |version, uuid|
       unless version == @@latest_version #バージョンが最新でなければ400　本当は転送したい
         status 400
+        body "invalid version. latest version is '#{@@latest_version}'"
       else
         save_dir = "#{@@data_dir}/#{uuid[0..1]}/#{uuid}"
         status_file_path = "#{save_dir}/status.json"
-        #ファイルがなければ400番?
-        status_json = JSON.parse(File.open(status_file_path).read)
-        response.headers["Access-Control-Allow-Origin"] = "*"
-        content_type :json
-        status_json.to_json
+        if File.exist?(status_file_path)
+          status_json = JSON.parse(File.open(status_file_path).read)
+          content_type :json
+          status_json.to_json
+        else
+          status 400
+          body "Invalid uuid"
+        end
       end
     end
 
     get '/api/:version/validation/:uuid/:filetype' do |version, uuid, filetype|
       unless version == @@latest_version #バージョンが最新でなければ400　本当は転送したい
         status 400
+        body "invalid version. latest version is '#{@@latest_version}'"
       else
         save_dir = "#{@@data_dir}/#{uuid[0..1]}/#{uuid}"
         file_list = Dir.glob("#{save_dir}/#{filetype}/*")
         if file_list.size == 1
           file_name = File.basename(file_list.first)
           file_path = file_list.first
-          response.headers["Access-Control-Allow-Origin"] = "*"
           send_file file_path, :filename => file_name, :type => 'application/xml'
         else
-          status 404
+          status 400
+          body "Invalid uuid or filetype"
         end
       end
     end
@@ -151,6 +160,7 @@ module DDBJValidator
     get '/api/:version/validation/:uuid/:filetype/autocorrect' do |version, uuid, filetype|
       unless version == @@latest_version #バージョンが最新でなければ400　本当は転送したい
         status 400
+        body "invalid version. latest version is '#{@@latest_version}'"
       else
         save_dir = "#{@@data_dir}/#{uuid[0..1]}/#{uuid}"
         result_file = "#{save_dir}/result.json"
@@ -164,11 +174,10 @@ module DDBJValidator
           AutoAnnotation.new().create_annotated_file(org_file, result_file, annotated_file_path, filetype)
         end
         if File.exist?(annotated_file_path)
-          response.headers["Access-Control-Allow-Origin"] = "*"
           send_file annotated_file_path, :filename => annotated_file_name, :type => 'application/xml'
         else
-          response.headers["Access-Control-Allow-Origin"] = "*"
-          status 404
+          status 400
+          body "Invalid uuid or filetype, or the auto-correct data is not exist of the uuid specified"
         end
       end
     end
